% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{ack}
\alias{ack}
\title{Acknowledge that the work on a message has finished successfully}
\usage{
ack(message)
}
\arguments{
\item{message}{The message object.}
}
\description{
Acknowledge that the work on a message has finished successfully
}
