% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{consume}
\alias{consume}
\title{Consume a message from a queue}
\usage{
consume(queue)
}
\arguments{
\item{queue}{The queue object.}
}
\value{
A message.
}
\description{
Blocks and waits for a message if there isn't one to work on currently.
}
