% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{create_queue}
\alias{create_queue}
\title{Create a queue in a database}
\usage{
create_queue(name = NULL, db = default_db())
}
\arguments{
\item{name}{Name of the queue. If not specified or \code{NULL}, a
name is generated randomly.}

\item{db}{Path to the database file.}
}
\description{
It also creates the database, if it does not exist.
}
