% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_ack}
\alias{db_ack}
\title{Positive or negative ackowledgement}
\usage{
db_ack(db, queue, id, lock, success)
}
\arguments{
\item{db}{DB file.}

\item{queue}{Queue name.}

\item{id}{Message id.}

\item{lock}{Name of the message lock file.}

\item{success}{Whether this is a positive or negative ACK.}
}
\description{
If positive, then we need to remove the message from the queue.
If negative, we just set the status to \code{FAILED}.
}
\keyword{internal}
