% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_consume}
\alias{db_consume}
\title{Consume a message from a message queue}
\usage{
db_consume(db, queue, poll_interval = 500)
}
\arguments{
\item{queue}{The queue object.}
}
\description{
This is the blocking version of \code{\link[=try_consume]{try_consume()}}. Currently it just
polls twice a second, and sleeps between the polls. Each poll will also
trigger a crash cleanup, if there are workers running.
}
\keyword{internal}
