% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_create_queue}
\alias{db_create_queue}
\title{Create a queue}
\usage{
db_create_queue(name, db)
}
\description{
The database columns:
\itemize{
\item id Id of the message, it is generated automatically by the database.
\item title The title of the message, can be empty. In the future,
it can be used to filter messages.
\item message The message, arbitrary text, can be empty.
\item status Can be:
\itemize{
\item \code{READY}, ready to be consumed
\item \code{WORKING}, it is being consumed
\item \code{FAILED}, failed.
}
}
}
\keyword{internal}
