% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{default_db}
\alias{default_db}
\title{The name of the default database}
\usage{
default_db()
}
\value{
A characater scalar, the name of the default database.
}
\description{
If the queue database is not specified explicitly,
then \code{liteq} uses this file. Its location is determined via the
\code{rappdirs} package, see \code{\link[rappdirs:user_data_dir]{rappdirs::user_data_dir()}}.
}
