% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{delete_queue}
\alias{delete_queue}
\title{Delete a queue}
\usage{
delete_queue(queue, force = FALSE)
}
\arguments{
\item{queue}{The queue to delete.}

\item{force}{Whether to delete the queue even if it contains messages.}
}
\description{
Delete a queue
}
