% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{ensure_db}
\alias{ensure_db}
\title{Ensure that the DB exists and has the right columns}
\usage{
ensure_db(db)
}
\arguments{
\item{db}{DB file.}
}
\description{
We try a query, and if it fails then we try to create the DB.
}
\keyword{internal}
