% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{ensure_queue}
\alias{ensure_queue}
\title{Make sure that a queue exists}
\usage{
ensure_queue(name, db = default_db())
}
\arguments{
\item{name}{Name of the queue.}

\item{db}{Path to the database file.}
}
\value{
The queue object.
}
\description{
If it does not exist, then the queue will be created.
}
