% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{is_empty}
\alias{is_empty}
\title{Check if a queue is empty}
\usage{
is_empty(queue)
}
\arguments{
\item{queue}{The queue object.}
}
\value{
Logical, whether the queue is empty.
}
\description{
Check if a queue is empty
}
\seealso{
\link{liteq} for examples

Other liteq messages: \code{\link{ack}},
  \code{\link{consume}},
  \code{\link{list_failed_messages}},
  \code{\link{list_messages}}, \code{\link{message_count}},
  \code{\link{publish}},
  \code{\link{remove_failed_messages}},
  \code{\link{requeue_failed_messages}},
  \code{\link{try_consume}}
}
\concept{liteq messages}
