% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{list_messages}
\alias{list_messages}
\title{List all messages in a queue}
\usage{
list_messages(queue)
}
\arguments{
\item{queue}{The queue object.}
}
\value{
Data frame with columns: \code{id}, \code{title}, \code{status}.
}
\description{
List all messages in a queue
}
