% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{list_queues}
\alias{list_queues}
\title{List all queues in a database}
\usage{
list_queues(db = default_db())
}
\arguments{
\item{db}{The queue database to query.}
}
\value{
A list of \code{liteq_queue} objects.
}
\description{
List all queues in a database
}
