% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liteq-package.R
\docType{package}
\name{liteq-package}
\alias{liteq-package}
\title{liteq: Lightweight Portable Message Queue Using 'SQLite'}
\description{
Temporary and permanent message queues for R. Built on top of 'SQLite' databases. 'SQLite' provides locking, and makes it possible to detect crashed consumers. Crashed jobs can be automatically marked as "failed", or put in the queue again, potentially a limited number of times.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/liteq#readme}
  \item Report bugs at \url{https://github.com/r-lib/liteq/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
