% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{liteq}
\alias{liteq}
\alias{liteq-package}
\title{Lightweight Portable Message Queue Using 'SQLite'}
\description{
Temporary and permanent message queues for R. Built on top of 'SQLite'
databases.
}

