% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{make_message}
\alias{make_message}
\title{Make a message object}
\usage{
make_message(id, title, message, db, queue, lockdir)
}
\arguments{
\item{id}{Message id, integer, auto-generated.}

\item{title}{Title of message.}

\item{message}{The message itself.}

\item{db}{Main DB file.}

\item{queue}{Name of the queue.}

\item{lockdir}{Directory to create the message lock in.}
}
\value{
message object
}
\description{
It creates the lock for the message as well.
}
\details{
The message object contains the connection to the message lock. If the
worker crashes, then there will be no reference to the connection, and
the lock will be released. This is how we detect crashed workers.
}
\keyword{internal}
