% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{message_count}
\alias{message_count}
\title{Get the number of messages in a queue.}
\usage{
message_count(queue)
}
\arguments{
\item{queue}{The queue object.}
}
\value{
Number of messages in the queue.
}
\description{
Get the number of messages in a queue.
}
\seealso{
\link{liteq} for examples

Other liteq messages: \code{\link{ack}},
  \code{\link{consume}}, \code{\link{is_empty}},
  \code{\link{list_failed_messages}},
  \code{\link{list_messages}}, \code{\link{publish}},
  \code{\link{remove_failed_messages}},
  \code{\link{requeue_failed_messages}},
  \code{\link{try_consume}}
}
\concept{liteq messages}
