% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{nack}
\alias{nack}
\title{Report that the work on a message has failed}
\usage{
nack(message)
}
\arguments{
\item{message}{The message object.}
}
\description{
Report that the work on a message has failed
}
\seealso{
\link{liteq} for examples
}
