% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{remove_failed_messages}
\alias{remove_failed_messages}
\title{Remove failed messages from the queue}
\usage{
remove_failed_messages(queue, id = NULL)
}
\arguments{
\item{queue}{The queue object.}

\item{id}{Ids of the messages to requeue. If it is \code{NULL}, then all
failed messages will be removed.}
}
\description{
Remove failed messages from the queue
}
