% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{requeue_failed_messages}
\alias{requeue_failed_messages}
\title{Requeue failed messages}
\usage{
requeue_failed_messages(queue, id = NULL)
}
\arguments{
\item{queue}{The queue object.}

\item{id}{Ids of the messages to requeue. If it is \code{NULL}, then all
failed messages will be requeued.}
}
\value{
The list of the requeued messages.
}
\description{
Requeue failed messages
}
