% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{ack}
\alias{ack}
\title{Acknowledge that the work on a message has finished successfully}
\usage{
ack(message)
}
\arguments{
\item{message}{The message object.}
}
\description{
Acknowledge that the work on a message has finished successfully
}
\seealso{
\link{liteq} for examples

Other liteq messages: \code{\link{consume}},
  \code{\link{list_failed_messages}},
  \code{\link{list_messages}}, \code{\link{publish}},
  \code{\link{remove_failed_messages}},
  \code{\link{requeue_failed_messages}},
  \code{\link{try_consume}}
}
