% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{consume}
\alias{consume}
\title{Consume a message from a queue}
\usage{
consume(queue)
}
\arguments{
\item{queue}{The queue object.}
}
\value{
A message.
}
\description{
Blocks and waits for a message if there isn't one to work on currently.
}
\seealso{
\link{liteq} for examples

Other liteq messages: \code{\link{ack}},
  \code{\link{list_failed_messages}},
  \code{\link{list_messages}}, \code{\link{publish}},
  \code{\link{remove_failed_messages}},
  \code{\link{requeue_failed_messages}},
  \code{\link{try_consume}}
}
