% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{publish}
\alias{publish}
\title{Publish a message in a queue}
\usage{
publish(queue, title = "", message = "")
}
\arguments{
\item{queue}{The queue object.}

\item{title}{The title of the message. It can be the empty string.}

\item{message}{The body of the message. It can be the empty string.}
}
\description{
Publish a message in a queue
}
