% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{remove_failed_messages}
\alias{remove_failed_messages}
\title{Remove failed messages from the queue}
\usage{
remove_failed_messages(queue, id = NULL)
}
\arguments{
\item{queue}{The queue object.}

\item{id}{Ids of the messages to requeue. If it is \code{NULL}, then all
failed messages will be removed.}
}
\description{
Remove failed messages from the queue
}
\seealso{
\link{liteq} for examples

Other liteq messages: \code{\link{ack}},
  \code{\link{consume}},
  \code{\link{list_failed_messages}},
  \code{\link{list_messages}}, \code{\link{publish}},
  \code{\link{requeue_failed_messages}},
  \code{\link{try_consume}}
}
