% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{requeue_failed_messages}
\alias{requeue_failed_messages}
\title{Requeue failed messages}
\usage{
requeue_failed_messages(queue, id = NULL)
}
\arguments{
\item{queue}{The queue object.}

\item{id}{Ids of the messages to requeue. If it is \code{NULL}, then all
failed messages will be requeued.}
}
\description{
Requeue failed messages
}
\seealso{
\link{liteq} for examples

Other liteq messages: \code{\link{ack}},
  \code{\link{consume}},
  \code{\link{list_failed_messages}},
  \code{\link{list_messages}}, \code{\link{publish}},
  \code{\link{remove_failed_messages}},
  \code{\link{try_consume}}
}
