% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{try_consume}
\alias{try_consume}
\title{Consume a message if there is one available}
\usage{
try_consume(queue)
}
\arguments{
\item{queue}{The queue object.}
}
\value{
A message, or \code{NULL} if there is not message to work on.
}
\description{
Consume a message if there is one available
}
\seealso{
\link{liteq} for examples

Other liteq messages: \code{\link{ack}},
  \code{\link{consume}},
  \code{\link{list_failed_messages}},
  \code{\link{list_messages}}, \code{\link{publish}},
  \code{\link{remove_failed_messages}},
  \code{\link{requeue_failed_messages}}
}
